pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lunar-seismic-fno"
version = "0.1.0"
description = "Companion code for waveform and spectrogram FNO seismic detection"
requires-python = ">=3.9"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "torch>=2.0",
  "scikit-learn>=1.3",
  "matplotlib>=3.7",
  "PyYAML>=6.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
